+++ commons-civil/src/main/resources/application.yml
@@ -1,0 +1,8 @@
+
+bankHolidays:
+  api:
+    url: https://www.gov.uk
+
+nonworking-days:
+  datafile: /data/non-working_days.dat
+
