+++ src/main/java/uk/gov/hmcts/reform/civil/service/WorkingDayIndicator.java
@@ -2,8 +2,8 @@ package uk.gov.hmcts.reform.civil.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import uk.gov.hmcts.reform.civil.service.bankholidays.NonWorkingDaysCollection;
-import uk.gov.hmcts.reform.civil.service.bankholidays.PublicHolidaysCollection;
+import uk.gov.hmcts.reform.bankholidays.NonWorkingDaysCollection;
+import uk.gov.hmcts.reform.bankholidays.PublicHolidaysCollection;
 
 import java.time.LocalDate;
